backend/requirements.txt
# ========================================
# WEB FRAMEWORK
# ========================================
fastapi==0.115.0
uvicorn[standard]==0.34.0

# ========================================
# AUTHENTICATION & SECURITY
# ========================================
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.12

# ========================================
# DATABASE
# ========================================
sqlalchemy==2.0.36
email-validator==2.2.0

# ========================================
# MISTRAL AI
# ========================================
mistralai==1.2.4
python-dotenv==1.0.1

# ========================================
# VOICE ANALYSIS (for voice-analysis-system)
# ========================================
librosa==0.10.2.post1
matplotlib==3.9.3
numpy==1.26.4
scipy==1.14.1
soundfile==0.12.1
audioread==3.0.1

# ========================================
# HTTP REQUESTS
# ========================================
requests==2.32.3
# httpx version controlled by mistralai package

# ========================================
# OPTIONAL: Local Ministral Model (HuggingFace)
# ========================================
# Uncomment if using local model instead of API
# transformers==4.46.0
# torch==2.5.1
# accelerate==1.2.1
